feat: add dashboard route and components for payment management
- Created a new dashboard route with a YearlyCarousel component for displaying payment data.
- Implemented authentication checks and redirection for logged-in users.
- Added loading and error handling states for better user experience.
- Refactored the index route to redirect authenticated users to the dashboard.
- Created a new login route with user-friendly UI and error handling.
